+++ inc/itil_project.class.php
@@ -137,8 +137,11 @@ class Itil_Project extends CommonDBRelation {
          $itemTable = $itemtype::getTable();
 
          $iterator = $DB->request([
-            'SELECT DISTINCT' => "$selfTable.id AS linkID",
-            'FIELDS'          => "$itemTable.*",
+            'SELECT'          => [
+               "$selfTable.id AS linkID",
+               "$itemTable.*"
+            ],
+            'DISTINCT'        => true,
             'FROM'            => $selfTable,
             'LEFT JOIN'       => [
                $itemTable => [
@@ -279,8 +282,11 @@ class Itil_Project extends CommonDBRelation {
       $projectTable = Project::getTable();
 
       $iterator = $DB->request([
-         'SELECT DISTINCT' => "$selfTable.id AS linkID",
-         'FIELDS'          => "$projectTable.*",
+         'SELECT'          => [
+            "$selfTable.id AS linkID",
+            "$projectTable.*"
+         ],
+         'DISTINCT'        => true,
          'FROM'            => $selfTable,
          'LEFT JOIN'       => [
             $projectTable => [