+++ inc/notificationtargetcommonitilobject.class.php
@@ -393,8 +393,11 @@ abstract class NotificationTargetCommonITILObject extends NotificationTarget {
          $fkfield           = $this->obj->getForeignKeyField();
 
          $iterator = $DB->request([
-            'SELECT DISTINCT' => 'glpi_suppliers.email AS email',
-            'FIELDS'          => 'glpi_suppliers.name AS name',
+            'SELECT'          => [
+               'glpi_suppliers.email AS email',
+               'glpi_suppliers.name AS name'
+            ],
+            'DISTINCT'        => true,
             'FROM'            => $supplierlinktable,
             'LEFT JOIN'       => [
                'glpi_suppliers'  => [